Removed cache setting
The caching duration is now hard coded into the widget.
Before the widget would cache the feed for the duration specified in
that instance of the widget and wouldn't reload it if another instance
of the widget had set a lower cache duration.
The cache duration is set at 10 minutes.